backend/pyproject.toml
[project]
name = "flood-risk-assessment-backend"
version = "0.1.0"
description = "FastAPI backend for flood risk assessment using Google AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "aiofiles>=24.1.0",
    "pillow>=11.3.0",
    "aiosqlite>=0.20.0",
    "requests>=2.31.0",
    "pyjwt>=2.8.0",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "httpx>=0.28.1",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
